pyproject.toml
[tool.poetry]
name = "mooc-cha-api"
version = "0.1.0"
description = ""
authors = ["Jaeyeol Lee <rijgndqw012@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.13"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.6.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"


[tool.aerich]
tortoise_orm = "core.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
